+++ spec/features/admin/accounts/index_page_spec.rb
@@ -94,8 +94,7 @@ describe "admin section" do
 
           context "is eligible to apply for cards but not ready" do
             let(:extra_setup) do
-              owner.update_attributes!(eligible: true)
-              owner.update_attributes!(ready: false)
+              owner.update_attributes!(eligible: true, ready: false)
             end
 
             it "has an 'E' by their name" do
@@ -105,8 +104,7 @@ describe "admin section" do
 
           context "is eligible and ready to apply for cards" do
             let(:extra_setup) do
-              owner.update_attributes!(eligible: true)
-              owner.update_attributes!(ready: true)
+              owner.update_attributes!(eligible: true, ready: true)
             end
 
             it "has an 'R' by their name" do
@@ -118,8 +116,7 @@ describe "admin section" do
         context "and has an eligible and ready companion" do
           let(:companion) { @accounts[2].companion }
           let(:extra_setup) do
-            companion.update_attributes!(eligible: true)
-            companion.update_attributes!(ready: true)
+            companion.update_attributes!(eligible: true, ready: true)
           end
 
           it "has an 'R' next to the companion's name" do